--- a/benchmark/bench-eltwise-fma-mod.cpp
+++ b/benchmark/bench-eltwise-fma-mod.cpp
@@ -17,75 +17,75 @@ namespace hexl {
 
 //=================================================================
 
-// state[0] is the degree
-static void BM_EltwiseFMAModNative(benchmark::State& state) {  //  NOLINT
+static void BM_EltwiseFMAModAddNative(benchmark::State& state) {  //  NOLINT
   size_t input_size = state.range(0);
   uint64_t modulus = 0xffffffffffc0001ULL;
+  bool add = state.range(1);
 
-  AlignedVector64<uint64_t> op1(input_size, 1);
-  uint64_t op2 = 1;
-  AlignedVector64<uint64_t> op3(input_size, 2);
+  AlignedVector64<uint64_t> input1(input_size, 1);
+  uint64_t input2 = 1;
+  AlignedVector64<uint64_t> input3(input_size, 2);
+  uint64_t* arg3 = add ? input3.data() : nullptr;
 
   for (auto _ : state) {
-    EltwiseFMAMod(op1.data(), op1.data(), op2, op3.data(), op1.size(), modulus,
-                  1);
+    EltwiseFMAMod(input1.data(), input1.data(), input2, arg3, input1.size(),
+                  modulus, 1);
   }
 }
 
-BENCHMARK(BM_EltwiseFMAModNative)
+BENCHMARK(BM_EltwiseFMAModAddNative)
     ->Unit(benchmark::kMicrosecond)
-    ->Args({1024})
-    ->Args({4096})
-    ->Args({16384});
+    ->ArgsProduct({{1024, 8192, 16384}, {false, true}});
 
 //=================================================================
 
 #ifdef HEXL_HAS_AVX512DQ
-// state[0] is the degree
 static void BM_EltwiseFMAModAVX512DQ(benchmark::State& state) {  //  NOLINT
   size_t input_size = state.range(0);
   size_t modulus = 100;
+  bool add = state.range(1);
 
   AlignedVector64<uint64_t> input1(input_size, 1);
   uint64_t input2 = 3;
   AlignedVector64<uint64_t> input3(input_size, 2);
 
+  uint64_t* arg3 = add ? input3.data() : nullptr;
+
   for (auto _ : state) {
-    EltwiseFMAModAVX512<64, 1>(input1.data(), input1.data(), input2,
-                               input3.data(), input_size, modulus);
+    EltwiseFMAModAVX512<64, 1>(input1.data(), input1.data(), input2, arg3,
+                               input_size, modulus);
   }
 }
 
 BENCHMARK(BM_EltwiseFMAModAVX512DQ)
     ->Unit(benchmark::kMicrosecond)
-    ->Args({1024})
-    ->Args({4096})
-    ->Args({16384});
+    ->ArgsProduct({{1024, 8192, 16384}, {false, true}});
 #endif
 
 //=================================================================
 
 #ifdef HEXL_HAS_AVX512IFMA
-// state[0] is the degree
 static void BM_EltwiseFMAModAVX512IFMA(benchmark::State& state) {  //  NOLINT
   size_t input_size = state.range(0);
   size_t modulus = 100;
+  bool add = state.range(1);
 
   AlignedVector64<uint64_t> input1(input_size, 1);
   uint64_t input2 = 3;
   AlignedVector64<uint64_t> input3(input_size, 2);
 
+  uint64_t* arg3 = add ? input3.data() : nullptr;
+
   for (auto _ : state) {
-    EltwiseFMAModAVX512<52, 1>(input1.data(), input1.data(), input2,
-                               input3.data(), input_size, modulus);
+    EltwiseFMAModAVX512<52, 1>(input1.data(), input1.data(), input2, arg3,
+                               input_size, modulus);
   }
 }
 
 BENCHMARK(BM_EltwiseFMAModAVX512IFMA)
     ->Unit(benchmark::kMicrosecond)
-    ->Args({1024})
-    ->Args({4096})
-    ->Args({16384});
+    ->ArgsProduct({{1024, 8192, 16384}, {false, true}});
+
 #endif
 
 }  // namespace hexl