--- a/hexl/eltwise/eltwise-fma-mod-avx512.cpp
+++ b/hexl/eltwise/eltwise-fma-mod-avx512.cpp
@@ -85,6 +85,7 @@ void EltwiseFMAModAVX512(uint64_t* result, const uint64_t* arg1, uint64_t arg2,
   __m512i varg2_barr = _mm512_set1_epi64(static_cast<int64_t>(arg2_barr));
 
   __m512i vmodulus = _mm512_set1_epi64(static_cast<int64_t>(modulus));
+  __m512i vneg_modulus = _mm512_set1_epi64(-static_cast<int64_t>(modulus));
   __m512i v2_modulus = _mm512_set1_epi64(static_cast<int64_t>(2 * modulus));
   __m512i v4_modulus = _mm512_set1_epi64(static_cast<int64_t>(4 * modulus));
   const __m512i* vp_arg1 = reinterpret_cast<const __m512i*>(arg1);
@@ -96,7 +97,7 @@ void EltwiseFMAModAVX512(uint64_t* result, const uint64_t* arg1, uint64_t arg2,
 
   if (arg3) {
     const __m512i* vp_arg3 = reinterpret_cast<const __m512i*>(arg3);
-    HEXL_LOOP_UNROLL_4
+    HEXL_LOOP_UNROLL_8
     for (size_t i = n / 8; i > 0; --i) {
       __m512i varg1 = _mm512_loadu_si512(vp_arg1);
       __m512i varg3 = _mm512_loadu_si512(vp_arg3);
@@ -106,16 +107,16 @@ void EltwiseFMAModAVX512(uint64_t* result, const uint64_t* arg1, uint64_t arg2,
       varg3 = _mm512_hexl_small_mod_epu64<InputModFactor>(
           varg3, vmodulus, &v2_modulus, &v4_modulus);
 
-      __m512i va_times_b = _mm512_hexl_mullo_epi<64>(varg1, varg2);
-
+      __m512i va_times_b = _mm512_hexl_mullo_epi<BitShift>(varg1, varg2);
       __m512i vq = _mm512_hexl_mulhi_epi<BitShift>(varg1, varg2_barr);
-      __m512i vq_times_mod = _mm512_mullo_epi64(vq, vmodulus);
-      vq = _mm512_sub_epi64(va_times_b, vq_times_mod);
-      // Conditional Barrett subtraction
-      vq = _mm512_hexl_small_mod_epu64(vq, vmodulus);
 
+      // Compute vq in [0, 2 * modulus)
+      vq = _mm512_hexl_mullo_add_lo_epi<BitShift>(va_times_b, vq, vneg_modulus);
+
+      // Add arg3, bringing vq to [0, 3 * modulus)
       vq = _mm512_add_epi64(vq, varg3);
-      vq = _mm512_hexl_small_mod_epu64(vq, vmodulus);
+      // Reduce to [0, modulus)
+      vq = _mm512_hexl_small_mod_epu64<4>(vq, vmodulus, &v2_modulus);
 
       _mm512_storeu_si512(vp_result, vq);
 
@@ -124,16 +125,16 @@ void EltwiseFMAModAVX512(uint64_t* result, const uint64_t* arg1, uint64_t arg2,
       ++vp_arg3;
     }
   } else {  // arg3 == nullptr
-    HEXL_LOOP_UNROLL_4
+    HEXL_LOOP_UNROLL_8
     for (size_t i = n / 8; i > 0; --i) {
       __m512i varg1 = _mm512_loadu_si512(vp_arg1);
       varg1 = _mm512_hexl_small_mod_epu64<InputModFactor>(
           varg1, vmodulus, &v2_modulus, &v4_modulus);
 
+      __m512i va_times_b = _mm512_hexl_mullo_epi<BitShift>(varg1, varg2);
       __m512i vq = _mm512_hexl_mulhi_epi<BitShift>(varg1, varg2_barr);
-      __m512i vq_times_mod = _mm512_mullo_epi64(vq, vmodulus);
-      __m512i va_times_b = _mm512_hexl_mullo_epi<64>(varg1, varg2);
-      vq = _mm512_sub_epi64(va_times_b, vq_times_mod);
+
+      vq = _mm512_hexl_mullo_add_lo_epi<BitShift>(va_times_b, vq, vneg_modulus);
       // Conditional Barrett subtraction
       vq = _mm512_hexl_small_mod_epu64(vq, vmodulus);
       _mm512_storeu_si512(vp_result, vq);