1.5x - 1.7x speedup on EltwiseFMAModAVX512IFMA using a few tricks also used in the NTT:

1) Switch to IFMA 52-bit mullo when computing `vq_times_mod`. The speedup is likely due to better pipelining on the same port

2) Use the negative modulus, `neg_p`. This helps compute `q := a * b - q * p` via two steps: `tmp = a * b;  q = fma(tmp, q, neg_p)` rather than the previous three steps: `tmp = a*b; tmp2 = q*p; q = tmp - tmp2` (see line 2 of Algorithm 4 in https://arxiv.org/pdf/2012.01968.pdf for background)

Additionally, for the case when the addition argument `arg3 != nullptr`, merging the two modular reductions led to a slight speedup.
 
On ICX with clang-12, I see
| Benchmark  | Before  | After   | Speedup  |
|---|---|---|---|
| BM_EltwiseFMAModAVX512IFMA/1024/0     | 0.196us   | 0.116us   | 1.69x   |   
| BM_EltwiseFMAModAVX512IFMA/8192/0     | 1.53us      | 1.03us   |  1.48x  |   
| BM_EltwiseFMAModAVX512IFMA/16384/0  | 3.05us     | 2.06us   |  1.51x |   
| BM_EltwiseFMAModAVX512IFMA/1024/1      | 0.268us  | 0.177us  |  1.51x |   
| BM_EltwiseFMAModAVX512IFMA/8192/1      | 2.22us     | 1.47us   |   1.51x |   
| BM_EltwiseFMAModAVX512IFMA/16384/1   | 4.43us    | 2.93us   |   1.51x |